Changed "script" to "command".
Forcing the use of a script is inconvenient when all you want to do is
run unix 'make'.  So, I changed it to talk about a "command" instead
of a "script", and made it skip checking for the existence of a script
if the argument doesn't have any path separators.