Need to grab the pid of the daemon process.
When configuring monit, it seems to be impossible to use just the
process name if the process forks a child.  (It's in the monit
documentation -- something about the child and parent temporarily both
having the same process name.)  So I changed things to use the pid.

This now outputs the pid when the -o flag is used.