Changed most uses of "which" to "that".
The differences between "which" and "that" are confusing.  A basic
rule of thumb is that "which" should be used to set off a
non-restrictive clause.  Restrictive clauses are usually easy to spot
because they are set off by commas.  If there is no comma-separated
clause in a sentence, then the use of "that" is usually called for; if
the clause is set apart using commas, then the use of "which" is
usually called for.  Here are two sources that explain this:

* http://www.getitwriteonline.com/archive/103103whichthat.htm
* http://www.quickanddirtytips.com/education/grammar/which-versus-that-0